requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1